requirements.txt
torch~=2.3.1+cu118
matplotlib~=3.8.4
numpy~=1.26.3
pillow~=10.2.0
torchvision~=0.18.1+cu118
tqdm~=4.66.4
pyyaml~=6.0.1
torchinfo~=1.8.0